Add BatchStart event
