+++ packages/keeper/package.json
@@ -19,6 +19,7 @@
   },
   "dependencies": {
     "@nomiclabs/hardhat-ethers": "2.0.2",
+    "bullmq": "1.51.3",
     "fastq": "1.13.0",
     "hardhat-dependency-compiler": "1.1.2"
   }
